refactor tape
