+++ python/taichi/__init__.py
@@ -4,6 +4,7 @@ from taichi._funcs import *
 from taichi._lib import core as _ti_core
 from taichi._logging import *
 from taichi._snode import *
+from taichi.ad import Tape
 from taichi.lang import *  # pylint: disable=W0622 # TODO(archibate): It's `taichi.lang.core` overriding `taichi.core`
 from taichi.types.annotations import *
 # Provide a shortcut to types since they're commonly used.
@@ -38,7 +39,8 @@ __deprecated_names__ = {
     'imshow': 'tools.imshow',
     'imwrite': 'tools.imwrite',
     'ext_arr': 'types.ndarray',
-    'any_arr': 'types.ndarray'
+    'any_arr': 'types.ndarray',
+    'Tape': 'ad.Tape'
 }
 
 __customized_deprecations__ = {
