+++ python/taichi/lang/impl.py
@@ -21,7 +21,7 @@ from taichi.lang.mesh import (ConvType, MeshElementFieldProxy, MeshInstance,
                               MeshReorderedScalarFieldProxy, element_type_name)
 from taichi.lang.snode import SNode
 from taichi.lang.struct import Struct, StructField, _IntermediateStruct
-from taichi.lang.tape import TapeImpl
+# from taichi.ad import Tape
 from taichi.lang.util import (cook_dtype, get_traceback, is_taichi_class,
                               python_scope, taichi_scope, warning)
 from taichi.types.primitive_types import f16, f32, f64, i32, i64, types
@@ -333,8 +333,9 @@ class PyTaichi:
         self._signal_handler_registry = None
         self.materialized = False
 
-    def get_tape(self, loss=None):
-        return TapeImpl(self, loss)
+    # def get_tape(self, loss=None):
+    #     from taichi.ad import Tape
+    #     return Tape(loss)
 
     def sync(self):
         self.materialize()
